+++ packages/secsync/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - added `onPendingChangesUpdated` callback
+- added `pendingChanges` config
 
 ### Fixed